multimodal_service/requirements.txt
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
pillow>=9.0.0
torch>=2.0.0
flask>=2.0.0
flask-cors>=3.0.0

